move Permit type definition to single place
